+++ releases/r1/microservices/embeddings/schemas/emebeddings.py
@@ -1,0 +1,11 @@
+from typing import List
+
+from pydantic import BaseModel
+
+
+class EmbeddingsRequest(BaseModel):
+    strings: List[str]
+
+
+class EmbeddingsResponse(EmbeddingsRequest):
+    embeddings: List[List[float]]
